+++ kinto_elasticsearch/indexer.py
@@ -9,18 +9,19 @@ logger = logging.getLogger(__name__)
 
 
 class Indexer(object):
-    def __init__(self, hosts):
+    def __init__(self, hosts, prefix="kinto"):
         self.client = elasticsearch.Elasticsearch(hosts)
+        self.prefix = prefix
 
     def search(self, bucket_id, collection_id, query, **kwargs):
-        indexname = '%s-%s' % (bucket_id, collection_id)
+        indexname = "{}-{}-{}".format(self.prefix, bucket_id, collection_id)
         return self.client.search(index=indexname,
                                   doc_type=indexname,
                                   body=query,
                                   **kwargs)
 
     def index_record(self, bucket_id, collection_id, record, id_field):
-        indexname = '%s-%s' % (bucket_id, collection_id)
+        indexname = "{}-{}-{}".format(self.prefix, bucket_id, collection_id)
         record_id = record[id_field]
 
         if not self.client.indices.exists(index=indexname):
@@ -34,13 +35,16 @@ class Indexer(object):
         return index
 
     def unindex_record(self, bucket_id, collection_id, record, id_field):
-        indexname = '%s-%s' % (bucket_id, collection_id)
+        indexname = "{}-{}-{}".format(self.prefix, bucket_id, collection_id)
         record_id = record[id_field]
         result = self.client.delete(index=indexname,
                                     doc_type=indexname,
                                     id=record_id,
                                     refresh=True)
         return result
+
+    def flush(self):
+        self.client.indices.delete(index="{}-*".format(self.prefix))
 
 
 def load_from_config(config):